Move test case from end of `stateful_pool` module to new `test_stateful_pool` test module.
